pyproject.toml
[project]
name = "aetherdock-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.32",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
]
